# xyzbooks/webapp/server/docs/swagger.yaml
basePath: /api
definitions:
  router.Author:
    properties:
      first_name:
        maxLength: 100
        type: string
      id:
        type: string
      last_name:
        maxLength: 100
        type: string
      middle_name:
        maxLength: 100
        type: string
    required:
    - first_name
    - last_name
    type: object
...
  router.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  router.ValidationErrors:
    properties:
      errors:
        items:
          $ref: '#/definitions/router.ValidationError'
        type: array
    type: object
info:
  contact: {}
  description: This is the API documentation for the XYZ Books API.
  title: XYZ Books API
  version: "1.0"
paths:
  /authors:
    get:
      consumes:
      - application/json
      description: Get all authors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/router.Author'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get all authors
      tags:
      - authors
    post:
      consumes:
      - application/json
      description: Create author
      parameters:
      - description: Author
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/router.Author'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.Author'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ValidationErrors'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create author
      tags:
      - authors
  /authors/{id}:
    delete:
      consumes:
      - application/json
      description: Delete author
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete author
      tags:
      - authors
    get:
      consumes:
      - application/json
      description: Get author by ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.Author'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get author by ID
      tags:
      - authors
    put:
      consumes:
      - application/json
      description: Update author
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      - description: Author
        in: body
        name: author
        required: true
        schema:
          $ref: '#/definitions/router.Author'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.Author'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ValidationErrors'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update author
      tags:
      - authors
...
swagger: "2.0"